reorganize into two modules
